% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aPretentiousJobTitleWords}
\alias{aPretentiousJobTitleWords}
\title{Pretentious job title words}
\format{
A list of list with named elements.
\describe{
  \item{names}{Language.}
}
}
\source{
\url{https://www.bullshitjob.com/title/}
}
\usage{
aPretentiousJobTitleWords
}
\description{
Pretentious job title words in different languages.
}
\keyword{datasets}
