% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfEnglishWords}
\alias{dfEnglishWords}
\title{English words data}
\format{
A data frame with 84996 rows and 4 columns.
\describe{
  \item{Word}{English word}
  \item{KnownWordQ}{Is the \code{Word} value a known word or not?}
  \item{CommonWordQ}{Is the \code{Word} value a common word or not?}
  \item{StopWordQ}{Is the \code{Word} value a stop word or not?}
}
}
\source{
\url{github.com}
}
\usage{
dfEnglishWords
}
\description{
English words data that correspond to that used in Mathematica's \code{RandomWord}.
There are 84923 words of type "Known", 39176 words of type "Common", 329 of type "Stopword".
}
\keyword{datasets}
