% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFunctions.R
\name{RandomPetName}
\alias{RandomPetName}
\title{Random pet name}
\usage{
RandomPetName(size = NULL, species = NULL, weighted = FALSE)
}
\arguments{
\item{size}{Number of pet names.}

\item{species}{Which species to use?
One of \code{c(NULL, "Cat", "Dog", "Goat", "Pig")}.
If NULL all species are used.}

\item{weighted}{Should the random choice of pet names be based on their popularity or not?}
}
\description{
Generates a vector of random pet names
}
\details{
See \url{https://catalog.data.gov/dataset/seattle-pet-licenses}.
}
