% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomDataFrameGenerator.R
\name{RandomString}
\alias{RandomString}
\title{Random strings.}
\usage{
RandomString(size = 1, lambda = 4)
}
\arguments{
\item{size}{Number of words.}

\item{lambda}{Poisson lambda for the string length distribution.}
}
\description{
Generates a vector of random strings.
}
