% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFunctions.R
\name{RandomPretentiousJobTitle}
\alias{RandomPretentiousJobTitle}
\title{Random pretentious job title}
\usage{
RandomPretentiousJobTitle(size = 1, numberOfWords = 3)
}
\arguments{
\item{size}{Number of job titles}

\item{numberOfWords}{Number of words per title.
One of \code{c(1, 2, 3, NULL)} or
a numerical vector of the values \code{c(1, 2, 3)}.}
}
\description{
Generates a vector of random POSIXct dates.
}
\details{
See https://www.bullshitjob.com/title/ .
}
